Rename ResumeListItem file
Rename `resume-li.tsx` to `resume-list-item.tsx`
